// web/src/api/model/bookCreateDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Rust Web App
 * OpenAPI spec version: 0.1.0
 */
import type { BookCreateDtoFormat } from './bookCreateDtoFormat';
import type { BookCreateDtoShopId } from './bookCreateDtoShopId';

export interface BookCreateDto {
  author: string;
  format?: BookCreateDtoFormat;
  price: number;
  publisher_id: string;
  shop_id?: BookCreateDtoShopId;
  title: string;
}